+++ Cargo.toml
@@ -28,7 +28,7 @@ indoc = "1.0"
 do-notation = "0.1.3"
 serde_json = "1.0"
 sha2 = "0.9.5"
-rand = "0.8.4"
+rand = "0.8.5"
 tempfile = "3.3.0"
 clvmr = "0.1.24"
 binascii = "0.1.4"
@@ -90,8 +90,3 @@ path = "src/classic/bins/shrink.rs"
 [[bin]]
 name = "repl"
 path = "src/classic/bins/repl.rs"
-
-[[bin]]
-name = "chia_prog_gen"
-path = "src/classic/bins/chia_prog_gen.rs"
-default-features = ["fuzzer"]
